Move parsing args and app_config to seperate module
